+++ tdclient/table_model.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 import time
 
-from tdclient._model import Model
+from tdclient.model import Model
 
 class Table(Model):
     """Database table on Treasure Data Service