+++ tdclient/test/account_model_test.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -16,7 +16,7 @@ def setup_function(function):
 
 def test_account():
     client = mock.MagicMock()
-    account = model.Account(client, 1, 2, storage_size=3456, guaranteed_cores=7, maximum_cores=8, created_at="created_at")
+    account = models.Account(client, 1, 2, storage_size=3456, guaranteed_cores=7, maximum_cores=8, created_at="created_at")
     assert account.account_id == 1
     assert account.plan == 2
     assert account.storage_size == 3456
@@ -26,11 +26,11 @@ def test_account():
 
 def test_account_storage_size_string():
     client = mock.MagicMock()
-    account1 = model.Account(client, 1, 1, storage_size=0)
+    account1 = models.Account(client, 1, 1, storage_size=0)
     assert account1.storage_size_string == "0.0 GB"
-    account2 = model.Account(client, 1, 1, storage_size=50*1024*1024)
+    account2 = models.Account(client, 1, 1, storage_size=50*1024*1024)
     assert account2.storage_size_string == "0.01 GB"
-    account3 = model.Account(client, 1, 1, storage_size=50*1024*1024*1024)
+    account3 = models.Account(client, 1, 1, storage_size=50*1024*1024*1024)
     assert account3.storage_size_string == "50.0 GB"
-    account4 = model.Account(client, 1, 1, storage_size=300*1024*1024*1024)
+    account4 = models.Account(client, 1, 1, storage_size=300*1024*1024*1024)
     assert account4.storage_size_string == "300 GB"