+++ tdclient/test/bulk_import_model_test.py
@@ -9,7 +9,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -17,7 +17,7 @@ def setup_function(function):
 
 def test_bulk_import():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     assert bulk_import.name == "name"
     assert bulk_import.database == "database"
     assert bulk_import.table == "table"
@@ -31,13 +31,13 @@ def test_bulk_import():
 
 def test_bulk_import_delete():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.delete()
     client.delete_bulk_import.assert_called_with("name")
 
 def test_bulk_import_freeze():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.freeze()
     client.freeze_bulk_import.assert_called_with("name")
@@ -45,7 +45,7 @@ def test_bulk_import_freeze():
 
 def test_bulk_import_unfreeze():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.unfreeze()
     client.unfreeze_bulk_import.assert_called_with("name")
@@ -53,7 +53,7 @@ def test_bulk_import_unfreeze():
 
 def test_bulk_import_perfom():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.perform()
     client.perform_bulk_import.assert_called_with("name")
@@ -61,7 +61,7 @@ def test_bulk_import_perfom():
 
 def test_bulk_import_commit():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.commit()
     client.commit_bulk_import.assert_called_with("name")
@@ -69,7 +69,7 @@ def test_bulk_import_commit():
 
 def test_bulk_import_upload_part():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.upload_part("part", b"bytes", 5)
     client.bulk_import_upload_part.assert_called_with("name", "part", b"bytes", 5)
@@ -77,7 +77,7 @@ def test_bulk_import_upload_part():
 
 def test_bulk_import_upload_file():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     stream = io.BytesIO(b"")
     bulk_import.upload_file("part", "json", stream)
@@ -86,7 +86,7 @@ def test_bulk_import_upload_file():
 
 def test_bulk_import_delete_part():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.delete_part("part")
     client.bulk_import_delete_part.assert_called_with("name", "part")
@@ -94,7 +94,7 @@ def test_bulk_import_delete_part():
 
 def test_bulk_import_list_parts():
     client = mock.MagicMock()
-    bulk_import = model.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
+    bulk_import = models.BulkImport(client, name="name", database="database", table="table", status="status", upload_frozen="upload_frozen", job_id="job_id", valid_records="valid_records", error_records="error_records", valid_parts="valid_parts", error_parts="error_parts")
     bulk_import.update = mock.MagicMock()
     bulk_import.list_parts()
     client.list_bulk_import_parts.assert_called_with("name")