+++ tdclient/test/database_model_test.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -16,7 +16,7 @@ def setup_function(function):
 
 def test_database():
     client = mock.MagicMock()
-    database = model.Database(client, "sample_datasets", tables=["nasdaq", "www_access"], count=12345, created_at="created_at", updated_at="updated_at", org_name="org_name", permission="administrator")
+    database = models.Database(client, "sample_datasets", tables=["nasdaq", "www_access"], count=12345, created_at="created_at", updated_at="updated_at", org_name="org_name", permission="administrator")
     assert database.org_name == "org_name"
     assert database.permission == "administrator"
     assert database.count == 12345
@@ -28,11 +28,11 @@ def test_database():
 def test_database_update_tables():
     client = mock.MagicMock()
     client.tables = mock.MagicMock(return_value=[
-        model.Table(client, "sample_datasets", "foo", "type", "schema", "count"),
-        model.Table(client, "sample_datasets", "bar", "type", "schema", "count"),
-        model.Table(client, "sample_datasets", "baz", "type", "schema", "count"),
+        models.Table(client, "sample_datasets", "foo", "type", "schema", "count"),
+        models.Table(client, "sample_datasets", "bar", "type", "schema", "count"),
+        models.Table(client, "sample_datasets", "baz", "type", "schema", "count"),
     ])
-    database = model.Database(client, "sample_datasets", tables=None, count=12345, created_at="created_at", updated_at="updated_at", org_name="org_name", permission="administrator")
+    database = models.Database(client, "sample_datasets", tables=None, count=12345, created_at="created_at", updated_at="updated_at", org_name="org_name", permission="administrator")
     tables = database.tables()
     assert [ table.name for table in tables ] == ["foo", "bar", "baz"]
     client.tables.assert_called_with("sample_datasets")