+++ tdclient/test/result_model_test.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -16,7 +16,7 @@ def setup_function(function):
 
 def test_result():
     client = mock.MagicMock()
-    result = model.Result(client, "name", "url", "org_name")
+    result = models.Result(client, "name", "url", "org_name")
     assert result.name == "name"
     assert result.url == "url"
     assert result.org_name == "org_name"