+++ tdclient/test/table_model_test.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -16,7 +16,7 @@ def setup_function(function):
 
 def test_table():
     client = mock.MagicMock()
-    table = model.Table(client, "db_name", "table_name", "type", "schema", 12345, created_at="created_at", updated_at="updated_at", estimated_storage_size=67890, last_import="last_import", last_log_timestamp="last_log_timestamp", expire_days="expire_days", primary_key="primary_key", primary_key_type="primary_key_type")
+    table = models.Table(client, "db_name", "table_name", "type", "schema", 12345, created_at="created_at", updated_at="updated_at", estimated_storage_size=67890, last_import="last_import", last_log_timestamp="last_log_timestamp", expire_days="expire_days", primary_key="primary_key", primary_key_type="primary_key_type")
     assert table.type == "type"
     assert table.db_name == "db_name"
     assert table.table_name == "table_name"
@@ -36,18 +36,18 @@ def test_table():
 
 def test_table_permission():
     client = mock.MagicMock()
-    table = model.Table(client, "sample_datasets", "nasdaq", "log", "schema", 12345)
+    table = models.Table(client, "sample_datasets", "nasdaq", "log", "schema", 12345)
     client.database().permission = "permission"
     assert table.permission == "permission"
     client.database.assert_called_with("sample_datasets")
 
 def test_table_estimated_storage_size_string():
     client = mock.MagicMock()
-    table1 = model.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=1)
+    table1 = models.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=1)
     assert table1.estimated_storage_size_string == "0.0 GB"
-    table2 = model.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=50*1024*1024)
+    table2 = models.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=50*1024*1024)
     assert table2.estimated_storage_size_string == "0.01 GB"
-    table3 = model.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=50*1024*1024*1024)
+    table3 = models.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=50*1024*1024*1024)
     assert table3.estimated_storage_size_string == "50.0 GB"
-    table4 = model.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=300*1024*1024*1024)
+    table4 = models.Table(client, "db_name", "table_name", "type", "schema", 12345, estimated_storage_size=300*1024*1024*1024)
     assert table4.estimated_storage_size_string == "300 GB"