+++ tdclient/test/user_model_test.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     import mock
 
-from tdclient import model
+from tdclient import models
 from tdclient.test.test_helper import *
 
 def setup_function(function):
@@ -16,7 +16,7 @@ def setup_function(function):
 
 def test_user():
     client = mock.MagicMock()
-    user = model.User(client, "name", "org_name", ["role1", "role2"], "email")
+    user = models.User(client, "name", "org_name", ["role1", "role2"], "email")
     assert user.name == "name"
     assert user.org_name == "org_name"
     assert user.role_names == ["role1", "role2"]